Fraction code

+++ prepare/Learning03/Program.cs
@@ -53,7 +53,7 @@ public class Fraction
 
     public double GetDecimalValue()
     {
-        return _top/_bottom; 
+        return (double)_top / (double)_bottom; 
     }
